refactor(webui): organize backend helpers
