+++ nanobot/channels/websocket.py
@@ -43,13 +43,20 @@ from nanobot.utils.media_decode import (
     save_base64_data_url,
 )
 from nanobot.utils.subagent_channel_display import scrub_subagent_messages_for_channel
-from nanobot.utils.webui_sidebar_state import (
+from nanobot.webui.settings_api import (
+    WebUISettingsError,
+    settings_payload,
+    update_agent_settings,
+    update_provider_settings,
+    update_web_search_settings,
+)
+from nanobot.webui.sidebar_state import (
     read_webui_sidebar_state,
     write_webui_sidebar_state,
 )
-from nanobot.utils.webui_thread_disk import delete_webui_thread
-from nanobot.utils.webui_transcript import append_transcript_object, build_webui_thread_response
-from nanobot.utils.webui_turn_helpers import websocket_turn_wall_started_at
+from nanobot.webui.thread_disk import delete_webui_thread
+from nanobot.webui.transcript import append_transcript_object, build_webui_thread_response
+from nanobot.webui.turn_helpers import websocket_turn_wall_started_at
 
 if TYPE_CHECKING:
     from nanobot.session.manager import SessionManager
@@ -224,47 +231,6 @@ def _query_first(query: dict[str, list[str]], key: str) -> str | None:
     """Return the first value for *key*, or None."""
     values = query.get(key)
     return values[0] if values else None
-
-
-def _mask_secret_hint(secret: str | None) -> str | None:
-    if not secret:
-        return None
-    if len(secret) <= 8:
-        return "••••"
-    return f"{secret[:4]}••••{secret[-4:]}"
-
-
-def _provider_requires_api_key(spec: Any) -> bool:
-    if spec.backend == "azure_openai":
-        return True
-    if spec.is_local or spec.is_direct:
-        return False
-    return True
-
-
-def _provider_configured_for_settings(spec: Any, provider_config: Any) -> bool:
-    if _provider_requires_api_key(spec):
-        return bool(provider_config.api_key)
-    return bool(
-        provider_config.api_key
-        or provider_config.api_base
-        or getattr(provider_config, "region", None)
-        or getattr(provider_config, "profile", None)
-    )
-
-
-_WEB_SEARCH_PROVIDER_OPTIONS: tuple[dict[str, str], ...] = (
-    {"name": "duckduckgo", "label": "DuckDuckGo", "credential": "none"},
-    {"name": "brave", "label": "Brave Search", "credential": "api_key"},
-    {"name": "tavily", "label": "Tavily", "credential": "api_key"},
-    {"name": "searxng", "label": "SearXNG", "credential": "base_url"},
-    {"name": "jina", "label": "Jina", "credential": "api_key"},
-    {"name": "kagi", "label": "Kagi", "credential": "api_key"},
-    {"name": "olostep", "label": "Olostep", "credential": "api_key"},
-)
-_WEB_SEARCH_PROVIDER_BY_NAME = {
-    provider["name"]: provider for provider in _WEB_SEARCH_PROVIDER_OPTIONS
-}
 
 
 def _parse_inbound_payload(raw: str) -> str | None:
@@ -800,150 +766,10 @@ class WebSocketChannel(BaseChannel):
         ]
         return _http_json_response({"sessions": cleaned})
 
-    def _settings_payload(self, *, requires_restart: bool = False) -> dict[str, Any]:
-        from nanobot.config.loader import get_config_path, load_config
-        from nanobot.providers.registry import PROVIDERS, find_by_name
-
-        config = load_config()
-        defaults = config.agents.defaults
-        active_preset_name = defaults.model_preset or "default"
-        try:
-            effective_preset = config.resolve_preset()
-        except Exception:
-            effective_preset = config.resolve_default_preset()
-            active_preset_name = "default"
-        provider_name = (
-            config.get_provider_name(effective_preset.model, preset=effective_preset)
-            or effective_preset.provider
-        )
-        provider = config.get_provider(effective_preset.model, preset=effective_preset)
-        selected_provider = provider_name
-        if effective_preset.provider != "auto":
-            spec = find_by_name(effective_preset.provider)
-            selected_provider = spec.name if spec else provider_name
-        providers = []
-        for spec in PROVIDERS:
-            provider_config = getattr(config.providers, spec.name, None)
-            if provider_config is None or spec.is_oauth:
-                continue
-            providers.append(
-                {
-                    "name": spec.name,
-                    "label": spec.label,
-                    "configured": _provider_configured_for_settings(spec, provider_config),
-                    "api_key_required": _provider_requires_api_key(spec),
-                    "api_key_hint": _mask_secret_hint(provider_config.api_key),
-                    "api_base": provider_config.api_base,
-                    "default_api_base": spec.default_api_base or None,
-                }
-            )
-        search_config = config.tools.web.search
-        search_provider = (
-            search_config.provider
-            if search_config.provider in _WEB_SEARCH_PROVIDER_BY_NAME
-            else "duckduckgo"
-        )
-        model_presets = [
-            {
-                "name": "default",
-                "label": "Default",
-                "active": active_preset_name == "default",
-                "is_default": True,
-                "model": defaults.model,
-                "provider": defaults.provider,
-                "max_tokens": defaults.max_tokens,
-                "context_window_tokens": defaults.context_window_tokens,
-                "temperature": defaults.temperature,
-                "reasoning_effort": defaults.reasoning_effort,
-            }
-        ]
-        for name, preset in config.model_presets.items():
-            model_presets.append(
-                {
-                    "name": name,
-                    "label": name,
-                    "active": active_preset_name == name,
-                    "is_default": False,
-                    "model": preset.model,
-                    "provider": preset.provider,
-                    "max_tokens": preset.max_tokens,
-                    "context_window_tokens": preset.context_window_tokens,
-                    "temperature": preset.temperature,
-                    "reasoning_effort": preset.reasoning_effort,
-                }
-            )
-        exec_config = config.tools.exec
-        return {
-            "agent": {
-                "model": effective_preset.model,
-                "provider": selected_provider,
-                "resolved_provider": provider_name,
-                "has_api_key": bool(provider and provider.api_key),
-                "model_preset": active_preset_name,
-                "max_tokens": effective_preset.max_tokens,
-                "context_window_tokens": effective_preset.context_window_tokens,
-                "temperature": effective_preset.temperature,
-                "reasoning_effort": effective_preset.reasoning_effort,
-                "timezone": defaults.timezone,
-                "bot_name": defaults.bot_name,
-                "bot_icon": defaults.bot_icon,
-                "tool_hint_max_length": defaults.tool_hint_max_length,
-            },
-            "model_presets": model_presets,
-            "providers": providers,
-            "web_search": {
-                "provider": search_provider,
-                "api_key_hint": _mask_secret_hint(search_config.api_key),
-                "base_url": search_config.base_url or None,
-                "max_results": search_config.max_results,
-                "timeout": search_config.timeout,
-                "providers": list(_WEB_SEARCH_PROVIDER_OPTIONS),
-            },
-            "web": {
-                "enable": config.tools.web.enable,
-                "proxy": config.tools.web.proxy,
-                "user_agent": config.tools.web.user_agent,
-                "search": {
-                    "max_results": search_config.max_results,
-                    "timeout": search_config.timeout,
-                },
-                "fetch": {
-                    "use_jina_reader": config.tools.web.fetch.use_jina_reader,
-                },
-            },
-            "runtime": {
-                "config_path": str(get_config_path().expanduser()),
-                "workspace_path": str(config.workspace_path),
-                "gateway_host": config.gateway.host,
-                "gateway_port": config.gateway.port,
-                "heartbeat": {
-                    "enabled": config.gateway.heartbeat.enabled,
-                    "interval_s": config.gateway.heartbeat.interval_s,
-                    "keep_recent_messages": config.gateway.heartbeat.keep_recent_messages,
-                },
-                "dream": {
-                    "schedule": defaults.dream.describe_schedule(),
-                    "max_batch_size": defaults.dream.max_batch_size,
-                    "max_iterations": defaults.dream.max_iterations,
-                    "annotate_line_ages": defaults.dream.annotate_line_ages,
-                },
-                "unified_session": defaults.unified_session,
-            },
-            "advanced": {
-                "restrict_to_workspace": config.tools.restrict_to_workspace,
-                "ssrf_whitelist_count": len(config.tools.ssrf_whitelist),
-                "mcp_server_count": len(config.tools.mcp_servers),
-                "exec_enabled": exec_config.enable,
-                "exec_sandbox": exec_config.sandbox or None,
-                "exec_path_append_set": bool(exec_config.path_append),
-            },
-            "requires_restart": requires_restart,
-        }
-
     def _handle_settings(self, request: WsRequest) -> Response:
         if not self._check_api_token(request):
             return _http_error(401, "Unauthorized")
-        return _http_json_response(self._settings_payload())
+        return _http_json_response(settings_payload())
 
     def _handle_commands(self, request: WsRequest) -> Response:
         if not self._check_api_token(request):
@@ -980,245 +806,32 @@ class WebSocketChannel(BaseChannel):
     def _handle_settings_update(self, request: WsRequest) -> Response:
         if not self._check_api_token(request):
             return _http_error(401, "Unauthorized")
-        from nanobot.config.loader import load_config, save_config
-        from nanobot.providers.registry import find_by_name
-
         query = _parse_query(request.path)
-        config = load_config()
-        defaults = config.agents.defaults
-        changed = False
-
-        if "model_preset" in query or "modelPreset" in query:
-            preset = _query_first(query, "model_preset")
-            if preset is None:
-                preset = _query_first(query, "modelPreset")
-            preset = (preset or "").strip()
-            preset_value = None if not preset or preset == "default" else preset
-            if preset_value is not None and preset_value not in config.model_presets:
-                return _http_error(400, "unknown model preset")
-            if defaults.model_preset != preset_value:
-                defaults.model_preset = preset_value
-                changed = True
-
-        model = _query_first(query, "model")
-        if model is not None:
-            model = model.strip()
-            if not model:
-                return _http_error(400, "model is required")
-            if defaults.model != model:
-                defaults.model = model
-                changed = True
-
-        provider = _query_first(query, "provider")
-        if provider is not None:
-            provider = provider.strip()
-            if not provider:
-                return _http_error(400, "provider is required")
-            if find_by_name(provider) is None:
-                return _http_error(400, "unknown provider")
-            provider_config = getattr(config.providers, provider, None)
-            spec = find_by_name(provider)
-            if (
-                provider_config is None
-                or spec is None
-                or not _provider_configured_for_settings(spec, provider_config)
-            ):
-                return _http_error(400, "provider is not configured")
-            if defaults.provider != provider:
-                defaults.provider = provider
-                changed = True
-
-        timezone = _query_first(query, "timezone")
-        if timezone is not None:
-            timezone = timezone.strip()
-            if not timezone:
-                return _http_error(400, "timezone is required")
-            try:
-                from zoneinfo import ZoneInfo
-                ZoneInfo(timezone)
-            except Exception:
-                return _http_error(400, "invalid timezone")
-            if defaults.timezone != timezone:
-                defaults.timezone = timezone
-                changed = True
-
-        bot_name = _query_first(query, "bot_name")
-        if bot_name is None:
-            bot_name = _query_first(query, "botName")
-        if bot_name is not None:
-            bot_name = bot_name.strip()
-            if not bot_name:
-                return _http_error(400, "bot_name is required")
-            if defaults.bot_name != bot_name:
-                defaults.bot_name = bot_name
-                changed = True
-
-        bot_icon = _query_first(query, "bot_icon")
-        if bot_icon is None:
-            bot_icon = _query_first(query, "botIcon")
-        if bot_icon is not None:
-            bot_icon = bot_icon.strip()
-            if defaults.bot_icon != bot_icon:
-                defaults.bot_icon = bot_icon
-                changed = True
-
-        tool_hint_max_length = _query_first(query, "tool_hint_max_length")
-        if tool_hint_max_length is None:
-            tool_hint_max_length = _query_first(query, "toolHintMaxLength")
-        if tool_hint_max_length is not None:
-            try:
-                parsed = int(tool_hint_max_length)
-            except ValueError:
-                return _http_error(400, "tool_hint_max_length must be an integer")
-            if parsed < 20 or parsed > 500:
-                return _http_error(400, "tool_hint_max_length must be between 20 and 500")
-            if defaults.tool_hint_max_length != parsed:
-                defaults.tool_hint_max_length = parsed
-                changed = True
-
-        if changed:
-            save_config(config)
-        # LLM provider/model changes are hot-reloaded by AgentLoop before each
-        # new turn via the provider snapshot loader, so a restart is unnecessary.
-        return _http_json_response(self._settings_payload(requires_restart=False))
+        try:
+            payload = update_agent_settings(query)
+        except WebUISettingsError as e:
+            return _http_error(e.status, e.message)
+        return _http_json_response(payload)
 
     def _handle_settings_provider_update(self, request: WsRequest) -> Response:
         if not self._check_api_token(request):
             return _http_error(401, "Unauthorized")
-        from nanobot.config.loader import load_config, save_config
-        from nanobot.providers.registry import find_by_name
-
         query = _parse_query(request.path)
-        provider_name = (_query_first(query, "provider") or "").strip()
-        if not provider_name:
-            return _http_error(400, "provider is required")
-        spec = find_by_name(provider_name)
-        if spec is None or spec.is_oauth:
-            return _http_error(400, "unknown provider")
-
-        config = load_config()
-        provider_config = getattr(config.providers, spec.name, None)
-        if provider_config is None:
-            return _http_error(400, "unknown provider")
-
-        changed = False
-        if "api_key" in query or "apiKey" in query:
-            api_key = _query_first(query, "api_key")
-            if api_key is None:
-                api_key = _query_first(query, "apiKey")
-            api_key = (api_key or "").strip() or None
-            if provider_config.api_key != api_key:
-                provider_config.api_key = api_key
-                changed = True
-
-        if "api_base" in query or "apiBase" in query:
-            api_base = _query_first(query, "api_base")
-            if api_base is None:
-                api_base = _query_first(query, "apiBase")
-            api_base = (api_base or "").strip() or None
-            if provider_config.api_base != api_base:
-                provider_config.api_base = api_base
-                changed = True
-
-        if changed:
-            save_config(config)
-        # API key/base changes are picked up by the next provider snapshot refresh.
-        return _http_json_response(self._settings_payload(requires_restart=False))
+        try:
+            payload = update_provider_settings(query)
+        except WebUISettingsError as e:
+            return _http_error(e.status, e.message)
+        return _http_json_response(payload)
 
     def _handle_settings_web_search_update(self, request: WsRequest) -> Response:
         if not self._check_api_token(request):
             return _http_error(401, "Unauthorized")
-        from nanobot.config.loader import load_config, save_config
-
         query = _parse_query(request.path)
-        provider_name = (_query_first(query, "provider") or "").strip().lower()
-        provider_option = _WEB_SEARCH_PROVIDER_BY_NAME.get(provider_name)
-        if provider_option is None:
-            return _http_error(400, "unknown web search provider")
-
-        config = load_config()
-        search_config = config.tools.web.search
-        web_config = config.tools.web
-        previous_provider = search_config.provider
-        changed = False
-
-        def set_search_value(attr: str, value: object) -> None:
-            nonlocal changed
-            if getattr(search_config, attr) != value:
-                setattr(search_config, attr, value)
-                changed = True
-
-        def set_fetch_value(attr: str, value: object) -> None:
-            nonlocal changed
-            if getattr(web_config.fetch, attr) != value:
-                setattr(web_config.fetch, attr, value)
-                changed = True
-
-        if search_config.provider != provider_name:
-            search_config.provider = provider_name
-            changed = True
-
-        credential = provider_option["credential"]
-        if credential == "none":
-            set_search_value("api_key", "")
-            set_search_value("base_url", "")
-        elif credential == "base_url":
-            base_url = _query_first(query, "base_url")
-            if base_url is None:
-                base_url = _query_first(query, "baseUrl")
-            base_url = base_url.strip() if base_url is not None else None
-            if not base_url and previous_provider == provider_name and search_config.base_url:
-                base_url = search_config.base_url
-            if not base_url:
-                return _http_error(400, "base_url is required")
-            set_search_value("base_url", base_url)
-            set_search_value("api_key", "")
-        else:
-            api_key = _query_first(query, "api_key")
-            if api_key is None:
-                api_key = _query_first(query, "apiKey")
-            api_key = api_key.strip() if api_key is not None else None
-            if not api_key and previous_provider == provider_name and search_config.api_key:
-                api_key = search_config.api_key
-            if not api_key:
-                return _http_error(400, "api_key is required")
-            set_search_value("api_key", api_key)
-            set_search_value("base_url", "")
-
-        max_results = _query_first(query, "max_results")
-        if max_results is None:
-            max_results = _query_first(query, "maxResults")
-        if max_results is not None:
-            try:
-                parsed = int(max_results)
-            except ValueError:
-                return _http_error(400, "max_results must be an integer")
-            if parsed < 1 or parsed > 10:
-                return _http_error(400, "max_results must be between 1 and 10")
-            set_search_value("max_results", parsed)
-
-        timeout = _query_first(query, "timeout")
-        if timeout is not None:
-            try:
-                parsed_timeout = int(timeout)
-            except ValueError:
-                return _http_error(400, "timeout must be an integer")
-            if parsed_timeout < 1 or parsed_timeout > 120:
-                return _http_error(400, "timeout must be between 1 and 120")
-            set_search_value("timeout", parsed_timeout)
-
-        use_jina_reader = _query_first(query, "use_jina_reader")
-        if use_jina_reader is None:
-            use_jina_reader = _query_first(query, "useJinaReader")
-        if use_jina_reader is not None:
-            normalized = use_jina_reader.strip().lower()
-            if normalized not in {"1", "0", "true", "false", "yes", "no"}:
-                return _http_error(400, "use_jina_reader must be boolean")
-            set_fetch_value("use_jina_reader", normalized in {"1", "true", "yes"})
-
-        if changed:
-            save_config(config)
-        return _http_json_response(self._settings_payload(requires_restart=False))
+        try:
+            payload = update_web_search_settings(query)
+        except WebUISettingsError as e:
+            return _http_error(e.status, e.message)
+        return _http_json_response(payload)
 
     @staticmethod
     def _is_websocket_channel_session_key(key: str) -> bool:
