+++ nanobot/webui/__init__.py
@@ -1,0 +1,2 @@
+"""Backend helpers for the bundled WebUI surface."""
+
